Implement processor flags as it's own type.
MAke sure we can do bitwise ops on it. Mostly this is its own type so a good printable string can be made.

Also fix process_branch and run_interrupt as they had some logic wrong.

TODO: Get rid of the 3 bools for interrupt state and make it an enum state machine