Reimplement debugging.
As originally done doesn't interact well with cargo test under load.

Printing each line is fine for a few instructions but the ROM tests do millions of ops.

Cargo will buffer all of this and it slows down things considerably.

Instead provide a Debug type that keeps everything in a circular buffer. 128 should be enough for now.

Then provide a macro which does an assert and on failure dumps the buffer along with an error string.

Convert everywhere we had the simple debug printer to this new setup.