Add an i/o package for basic in/out pins and definitions for tying them high/low.
Use this to define the 6 i/o pins on a 6510. This happens via custom Memory which wraps the passed in one.

This then requires Memory references to now get boxed as simply using references won't work once they are layered.

Expose the I/O ports via a function on the CPU so memory banks can use that to see what should be mapped in.