Convert ram() in the Memory trait to take a mutable array ref in.
Then it can copy.

I tried making it just return the array but something then made debug() run 4x slower even if that copy wasn't performed. Acted as if the function grew too large for icache or something.

Get rid of all the Box::leak in tests now and remove that from the lsan suppressions (leave commented out for posterity)

Fix debug in the non dump case to properly handle a wrapped PC. We were waiting for a panic here otherwise.

Put the Copy/Clone back on Register due to clipy.