A bunch of changes which improve monitor performance a bunch. (#34)
StepN now takes the pre-allocated output vec instead of making one. This allows caller reuse.

Remove debug mode except when we need it.

Now we can get 128 samples with RAM snapshots in < 2ms on a SPR machine.

Expose op_tick() from CPU so we don't have to debug to know state.

* Fixes so sanitizer will work since the timeout is so wild.

This is nightly only so we can't feature enable this without converting the whole project to nightly.

Instead for the sanitizer test so ahead and edit the code to add the cfg options we need.