Extended_cpu_tests (#62)
* load_register should take a Wrapping if that's how we return regs too. Consistency

Investigate op_addr/op_val and it makes the code way messier in the end. In the few places we add just make sure we do it with Wrapping.

* Very large update:

Added 2 more CPUs:

Rockwell 65C02 - Doesn't have STP/WAI but otherwise same as WDC
65SC02 - Doesn't have BBR/BBS/RMB/SMB (replace with various NOPs).

In all the CMOE cases there's still disagreement on the bus cycles in some places. Opened one issue for one (BBR/BBS). The rest I just kept notes on for now.

Everything passes if we ignore the ones from above so works for what I need. Found a few actual bugs in the end.