Turn debugging into a boxed trait object instead of a direct FnMut
It's difficult to manage lifetimes of a ref to an FnMut along with other items such as a Box<dyn CPU>. This is what led to just leaking before...

This does mean potentially lots of cloning for a debug object and all internal state really needs to be wrapped in Rc<RefCell<X>>

No visible problems with speed.

Back out all the leak check skips for monitor.