Turns out I misread catch_unwind
It returns an Err for panic (and assert then) but not regular errors. Those get wrapped in Ok()

So we need to unwrap the entire set of possibilities here to check whether to run a 2nd time for the rom/coverage tests.

Also make sure it's apparent this ran twice.